+++ lib/tidewave/router.ex
@@ -21,11 +21,11 @@ defmodule Tidewave.Router do
     |> halt()
   end
 
-  get "/app" do
+  get "/connect" do
     conn
     |> put_resp_content_type("text/html")
     |> put_resp_header("content-security-policy", "base-uri 'self'; frame-ancestors 'self';")
-    |> send_resp(200, control_html())
+    |> send_resp(200, control_html(conn))
     |> halt()
   end
 
@@ -59,7 +59,7 @@ defmodule Tidewave.Router do
 
     conn
     |> Plug.Parsers.call(opts)
-    |> MCP.Server.handle_http_message()
+    |> MCP.HTTP.handle_message()
     |> halt()
   end
 
@@ -239,10 +239,6 @@ defmodule Tidewave.Router do
   defp entrypoint_html do
     client_url = Application.get_env(:tidewave, :client_url, "https://tidewave.ai")
 
-    # We return a basic page that is used by Tidewave Web.
-    # Note that, by itself, this page is harmless and it
-    # cannot invoke any of the MCP endpoints, since the MCP
-    # refuses any requests with an Origin header.
     """
     <html>
       <head>
@@ -255,7 +251,7 @@ defmodule Tidewave.Router do
     """
   end
 
-  defp control_html do
+  defp control_html(conn) do
     client_url = Application.get_env(:tidewave, :client_url, "https://tidewave.ai")
 
     """
@@ -263,6 +259,7 @@ defmodule Tidewave.Router do
       <head>
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <meta name="tidewave:config" content="#{Tidewave.tidewave_config_meta(conn) |> Jason.encode!() |> Plug.HTML.html_escape()}" />
         <script type="module" src="#{client_url}/tc/control.js"></script>
       </head>
       <body></body>